Determined the need to authenicate config page.
As it currently stands the config page does little to autheniticate that the user is
allowed to be doing what they are doing. And with this page planned to have config
changes that would allow you to route notifications I have concluded this will be
needed.

This commit contains the experimentation on how I might go about doing that.